+++ __tests__/context-and-provider.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { describe, it, expect } from 'vitest';
 import { render, act } from '@testing-library/react';
-import { createSlice, createContext, createProvider } from '../src';
+import { createSlice, createContext, createProvider, createHooks } from '../src';
 
 describe('Context and Provider', () => {
   it('initialization store and usage Context', () => {
@@ -9,9 +9,10 @@ describe('Context and Provider', () => {
     const Context = createContext();
     const slice = createSlice<Slice>(Context);
     const store = slice.createStore({ count: 1 });
+    const hooks = createHooks<Slice>(slice.sliceId, Context)
     let trigger!: () => number;
     const TestComponent = () => {
-      const store = slice.useStore();
+      const store = hooks.useStore();
       trigger = () => {
         // ! test it
         return store.get('count');
@@ -27,26 +28,29 @@ describe('Context and Provider', () => {
     expect(trigger()).toBe(1);
   });
 
-  it('multiple stores one provider', () => {
+  it('Multiple stores one provider', () => {
     const Context = createContext();
-    type Slice = { count: number };
-    const slice1 = createSlice<Slice>(Context);
-    const slice2 = createSlice<Slice>(Context);
+    type Slice1 = { count: number };
+    type Slice2 = { name: string };
+    const slice1 = createSlice<Slice1>();
+    const slice2 = createSlice<Slice2>();
 
     const data1 = { count: 1 };
-    const data2 = { count: 2 };
+    const data2 = { name: "test" };
     const store1 = slice1.createStore(data1);
     const store2 = slice2.createStore(data2);
+    const hooks1 = createHooks<Slice1>(slice1.sliceId, Context)
+    const hooks2 = createHooks<Slice2>(slice2.sliceId, Context)
     let trigger!: () => void;
     let countTest1 = 0;
-    let countTest2 = 0;
+    let countTest2 = "";
     const TestComponent = () => {
-      const store1 = slice1.useStore();
-      const store2 = slice2.useStore();
+      const store1 = hooks1.useStore();
+      const store2 = hooks2.useStore();
       trigger = () => {
         // ! test it
         countTest1 = store1.get('count');
-        countTest2 = store2.get('count');
+        countTest2 = store2.get('name');
       };
       return null;
     };
@@ -61,7 +65,7 @@ describe('Context and Provider', () => {
       trigger();
     });
     expect(countTest1).toBe(data1.count);
-    expect(countTest2).toBe(data2.count);
+    expect(countTest2).toBe(data2.name);
   });
 
   it('one store multiple providers', () => {
@@ -70,10 +74,11 @@ describe('Context and Provider', () => {
     const slice = createSlice<Slice>(Context);
 
     const store = slice.createStore({ count: 1 });
+    const hooks = createHooks<Slice>(slice.sliceId, Context)
     let countTest = 0;
 
     const TestComponent1 = () => {
-      const [count] = slice.useState('count');
+      const [count] = hooks.useState('count');
       React.useEffect(() => {
         countTest = count;
       }, [count]);
@@ -82,7 +87,7 @@ describe('Context and Provider', () => {
     let trigger!: () => void;
 
     const TestComponent2 = () => {
-      const store = slice.useStore();
+      const store = hooks.useStore();
       trigger = () => {
         // ! test it
         store.set('count', 2);
@@ -108,51 +113,4 @@ describe('Context and Provider', () => {
 
     expect(countTest).toBe(2);
   });
-
-  // it('one slice with one store and multiple contexts', () => {
-  //   type Slice = { count: number };
-  //   const slice = createSlice<Slice>(null);
-  //   const store = slice.createStore({ count: 1 });
-
-  //   const Context1 = createContext();
-  //   const Context2 = createContext();
-
-  //   let countTest = 0;
-  //   const TestComponent1 = () => {
-  //     const [count] = slice.useState('count', Context1);
-  //     React.useEffect(() => {
-  //       countTest = count;
-  //     }, [count]);
-  //     return null;
-  //   };
-  //   let trigger!: () => void;
-
-  //   const TestComponent2 = () => {
-  //     const store = slice.useStore(Context2);
-  //     trigger = () => {
-  //       // ! test it
-  //       store.set('count', 2);
-  //     };
-  //     return null;
-  //   };
-  //   const Provider1 = createProvider(Context1);
-  //   const Provider2 = createProvider(Context2);
-
-  //   render(
-  //     <>
-  //       <Provider1 value={[store]}>
-  //         <TestComponent1 />
-  //       </Provider1>
-  //       <Provider2 value={[store]}>
-  //         <TestComponent2 />
-  //       </Provider2>
-  //     </>,
-  //   );
-
-  //   act(() => {
-  //     trigger();
-  //   });
-
-  //   expect(countTest).toBe(2);
-  // });
 });