+++ __tests__/error.test.tsx
@@ -1,37 +1,31 @@
 import { describe, it, expect } from 'vitest';
 import { render } from '@testing-library/react';
-import { createSlice, createContext, createProvider } from '../src';
+import { createSlice, createContext, createProvider, createHooks } from '../src';
 import { formatError } from '../src/helpers/error';
 
 describe('Error', () => {
   it('Hook must be used within a React Provider.', () => {
     type Slice = { one: string };
     const Context = createContext();
-    const slice = createSlice<Slice>(null);
-    const store = slice.createStore({ one: '' });
-    const Provider = createProvider(Context);
+    const slice = createSlice<Slice>();
+    const hooks = createHooks<Slice>(slice.sliceId, Context);
     const Test = () => {
       // throw error need Context
-      slice.useStore();
+      hooks.useStore();
       return null;
     };
 
-    expect(() =>
-      render(
-        <Provider value={[store]}>
-          <Test />
-        </Provider>,
-      ),
-    ).toThrow(formatError['contextNotExist']('useStore', null));
+    expect(() => render(<Test />)).toThrow(formatError['hookMustBeInsideProvider']());
   });
 
   it('The storage does not exist in the React Provider.', () => {
     type Slice = { one: string };
     const Context = createContext();
     const slice = createSlice<Slice>(Context);
+    const hooks = createHooks<Slice>(slice.sliceId, Context);
     const Provider = createProvider(Context);
     const Test = () => {
-      slice.useStore();
+      hooks.useStore();
       return null;
     };
     expect(() =>
@@ -41,19 +35,20 @@ describe('Error', () => {
           <Test />
         </Provider>,
       ),
-    ).toThrow(formatError['storeNotExist']('useStore', null));
+    ).toThrow(formatError['storeNotExist']());
   });
 
   it('In the useAsync hook, only IAsync values can be used.', () => {
     type Slice = { one: string };
     const Context = createContext();
-    const slice = createSlice<Slice>(Context);
+    const slice = createSlice<Slice>("Context");
     const store = slice.createStore({ one: '' });
+    const hooks = createHooks<Slice>(slice.sliceId, Context);
     const Provider = createProvider(Context);
 
     const Test = () => {
       // prop "one" must be IAsync type
-      slice.useAsync('one', () => () => {});
+      hooks.useAsync('one', () => () => {});
       return null;
     };