+++ src/context.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
-import { IContext, IStore } from './types';
-
+import { Store } from './store';
+import { IContext } from './types';
+import { ISliceId } from './types';
+import { formatError } from './helpers/error';
 /**
  * Creates a Context.
  */
@@ -12,7 +14,7 @@ export const createContext = () => {
 /**
  * Creates a Provider
  *
- * The provider accepts multiple store instances and maps them by `uniqId`.
+ * The provider accepts multiple store instances and maps them by `sliceId`.
  * This allows different slices to coexist in the same React tree.
  *
  *  @param Context React Context
@@ -20,14 +22,18 @@ export const createContext = () => {
 export const createProvider = (Context: React.Context<IContext>) => {
   type Props = {
     children: React.ReactNode;
-    value: IStore<any>[];
+    value: Store<any>[];
   };
 
   const Provider = ({ children, value }: Props) => {
     const [context] = React.useState<IContext>(() => {
-      const map = new Map<{}, IStore<any>>();
+      const map = new Map<ISliceId, Store<any>>();
       for (const store of value) {
-        map.set(store.uniqId, store);
+        if (store.sliceId !== null) {
+          map.set(store.sliceId, store);
+        } else {
+          throw formatError['providerSliceIdNull']();
+        }
       }
 
       return map;