+++ src/hooks/hooks.tsx
@@ -4,14 +4,14 @@ import type { IReducers, IContext } from '../types';
 export type HookOf<T> = T extends { hook: (...args: any[]) => any } ? T['hook'] : never;
 
 export class Hooks<T extends object, R extends IReducers<T> = {}> {
-  constructor(uniqId: object, Context: React.Context<IContext>, reducers?: R) {
-    this.useAsync = new UseAsync<T>(uniqId, Context).hook;
+  constructor(sliceId: object, Context: React.Context<IContext>, reducers?: R) {
+    this.useAsync = new UseAsync<T>(sliceId, Context).hook;
 
-    this.useState = new UseState<T>(uniqId, Context).hook;
+    this.useState = new UseState<T>(sliceId, Context).hook;
 
-    this.useReducer = new UseReducer<T, R>(uniqId, Context, reducers).hook;
+    this.useReducer = new UseReducer<T, R>(sliceId, Context, reducers).hook;
 
-    this.useStore = new UseStore<T>(uniqId, Context).hook;
+    this.useStore = new UseStore<T>(sliceId, Context).hook;
   }
   useAsync: HookOf<UseAsync<T>>;
   useState: HookOf<UseState<T>>;
@@ -20,9 +20,9 @@ export class Hooks<T extends object, R extends IReducers<T> = {}> {
 }
 
 export const createHooks = <T extends object, R extends IReducers<T> = {}>(
-  uniqId: object,
+  sliceId: object,
   Context: React.Context<IContext>,
   reducers?: R,
 ) => {
-  return new Hooks<T, R>(uniqId, Context, reducers);
+  return new Hooks<T, R>(sliceId, Context, reducers);
 };