+++ src/types.ts
@@ -1,22 +1,10 @@
-import { type Store } from 'observable-store-light';
+import { Store } from './store';
 
 /**
  * Context value type.
- * key store unique `uniqId`, value store
+ * key store unique `sliceId`, value store
  */
-export type IContext = Map<{}, IStore<any>>;
-
-export type IStore<T extends object> = Store<T> & { uniqId: {} };
-
-export type ISubStore<T extends object> = {
-  /**
-   * unlink state
-   * 
-   * @param hard - default true, unlint and assign new undepended state to this store
-   * @returns
-   */
-  unLinkState: (hard?: boolean) => void;
-} & IStore<T>;
+export type IContext = Map<ISliceId, Store<any>>;
 
 /**
  * Reducer function.
@@ -24,9 +12,14 @@ export type ISubStore<T extends object> = {
  * The first argument is always the store instance,
  * the rest are user-defined arguments.
  */
-export type IReducer<T extends object> = (...args: any[]) => (store: IStore<T>) => void;
+export type IReducer<T extends object> = (...args: any[]) => (store: Store<T>) => void;
 
 /**
- * type from record reducers
+ * Type from record reducers
  */
 export type IReducers<T extends object> = Record<string, IReducer<T>>;
+
+/**
+ * Uniq id for slice
+ */
+export type ISliceId = {};